Update hostname
Instead of hard coding the hostname to "-", read it from spec.address
to match how it is done on Linux.

It's worth noting that on Windows, spec.address returns the internal IP
of the instance, whereas on Linux it returns a formatted string
including the instance ID and instance name. We believe the BOSH Agent
on Windows should be updated so that spec.address returns the same
format as Linux.

[#160740764]

Signed-off-by: Natalie Arellano <narellano@pivotal.io>